Applied patch #1096916 to support building with MSVC 8. 	* generic/regerror.c: Avoid use of reserved word. 	* generic/tcl.h:      Select the right Tcl_Stat structure 	* generic/tclDate.c:  Casts to handle 64 bit time_t case. 	* tests/env.test:     Include essential envvar on Win32 	* win/nmakehlp.c:     Handle new return codes. 	* win/makefile.vc:    Use the selected options. 	* win/rules.vc:       Check options are applicable 	* win/tclWinPort.h:   Disable deprecated function warnings 	* win/tclWinSock.c:   Provide default value to avoid warning. 	* win/tclWinTime.c:   Add casts to handle 64bit time_t type.

+++ generic/tclDate.c
@@ -2359,17 +2359,17 @@ TclClockOldscanObjCmd( clientData, interp, objc, objv )
     resultElement = Tcl_NewObj();
     if ( yyHaveDate ) {
 	Tcl_ListObjAppendElement( interp, resultElement,
-				  Tcl_NewIntObj( yyYear ) );
+				  Tcl_NewIntObj( (int) yyYear ) );
 	Tcl_ListObjAppendElement( interp, resultElement,
-				  Tcl_NewIntObj( yyMonth ) );
+				  Tcl_NewIntObj( (int) yyMonth ) );
 	Tcl_ListObjAppendElement( interp, resultElement,
-				  Tcl_NewIntObj( yyDay ) );
+				  Tcl_NewIntObj( (int) yyDay ) );
     }
     Tcl_ListObjAppendElement( interp, result, resultElement );
 
     if ( yyHaveTime ) {
 	Tcl_ListObjAppendElement( interp, result,
-				  Tcl_NewIntObj( ToSeconds( yyHour,
+				  Tcl_NewIntObj( (int) ToSeconds( yyHour,
 							    yyMinutes,
 							    yySeconds,
 							    yyMeridian ) ) );
@@ -2380,7 +2380,7 @@ TclClockOldscanObjCmd( clientData, interp, objc, objv )
     resultElement = Tcl_NewObj();
     if ( yyHaveZone ) {
 	Tcl_ListObjAppendElement( interp, resultElement,
-				  Tcl_NewIntObj( -yyTimezone ) );
+				  Tcl_NewIntObj( (int) -yyTimezone ) );
 	Tcl_ListObjAppendElement( interp, resultElement,
 				  Tcl_NewIntObj( 1-yyDSTmode ) );
     }
@@ -2389,29 +2389,29 @@ TclClockOldscanObjCmd( clientData, interp, objc, objv )
     resultElement = Tcl_NewObj();
     if ( yyHaveRel ) {
 	Tcl_ListObjAppendElement( interp, resultElement,
-				  Tcl_NewIntObj( yyRelMonth ) );
+				  Tcl_NewIntObj( (int) yyRelMonth ) );
 	Tcl_ListObjAppendElement( interp, resultElement,
-				  Tcl_NewIntObj( yyRelDay ) );
+				  Tcl_NewIntObj( (int) yyRelDay ) );
 	Tcl_ListObjAppendElement( interp, resultElement,
-				  Tcl_NewIntObj( yyRelSeconds ) );
+				  Tcl_NewIntObj( (int) yyRelSeconds ) );
     }
     Tcl_ListObjAppendElement( interp, result, resultElement );
 
     resultElement = Tcl_NewObj();
     if ( yyHaveDay && !yyHaveDate ) {
 	Tcl_ListObjAppendElement( interp, resultElement,
-				  Tcl_NewIntObj( yyDayOrdinal ) );
+				  Tcl_NewIntObj( (int) yyDayOrdinal ) );
 	Tcl_ListObjAppendElement( interp, resultElement,
-				  Tcl_NewIntObj( yyDayNumber ) );
+				  Tcl_NewIntObj( (int) yyDayNumber ) );
     }
     Tcl_ListObjAppendElement( interp, result, resultElement );
 
     resultElement = Tcl_NewObj();
     if ( yyHaveOrdinalMonth ) {
 	Tcl_ListObjAppendElement( interp, resultElement,
-				  Tcl_NewIntObj( yyMonthOrdinal ) );
+				  Tcl_NewIntObj( (int) yyMonthOrdinal ) );
 	Tcl_ListObjAppendElement( interp, resultElement,
-				  Tcl_NewIntObj( yyMonth ) );
+				  Tcl_NewIntObj( (int) yyMonth ) );
     }
     Tcl_ListObjAppendElement( interp, result, resultElement );
